\name{prueba1-package}
\alias{prueba1-package}
\alias{prueba1}
\docType{package}
\title{
\packageTitle{prueba1}
}
\description{
\packageDescription{prueba1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{prueba1}
\packageIndices{prueba1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{prueba1}

Maintainer: \packageMaintainer{prueba1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
